src/requirements.txt
flask
flasgger
python-dotenv
supabase
osmnx
gunicorn
googlemaps
geopandas
shapely
flask-cors
APScheduler==3.11.0
scikit-learn
psycopg2-binary